# documentation/swaggerFe.yaml
openapi: 3.0.3
info:
  title: Learn Wise API App
  description: |-
    lorem ipsum dolor is amet
  contact:
    email: mahadiocsidio@students.amikom.ac.id
  version: 1.0.0
servers:
  - url: https://learn-wise-develop.up.railway.app
  - url: https://binar-academy-final-project.vercel.app
  - url: http://localhost:3000
...
paths:
...
  /auth/verify-otp:
    post:
      tags:
        - Login User
      operationId: verify-otp
      summary: Verify OTP
      security:
        - bearerAuth: []
      requestBody:
        required: true
        $ref: "#/components/requestBodies/verify-otp"
      responses:
        '200':
          description: OTP verification successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Activation Code verified successfully'
                  err:
                    type: string
                    example: null
                  data:
                    type: object
                    properties:
                      email:
                        type: string
                        format: email
                        example: user@example.com
                      otp:
                        type: string
                        example: 123456
                      is_verified:
                        type: boolean
                        example: true
        '400':
          description: Failed OTP verification
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Bad Request'
                  error:
                    type: string
                    example: 'Invalid Email or OTP'
                  data:
                    example: null
...
  /auth/reset-password:
    post:
      tags:
        - No Login User
      operationId: reset-password
      summary: Reset Password
      requestBody:
        description: Reset password request
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
      responses:
        '200':
          description: Password reset email sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Send'
                  err:
                    type: string
                    example: null
                  data:
                    type: object
                    properties:
                      email:
                        type: string
                        format: email
                        example: user@example.com
        '400':
          description: Email not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Email Not Found'
                  err:
                    type: string
                    example: 'Enter Registered Email!'
                  data:
                    example: null
    put:
      tags:
        - No Login User
      operationId: change-password
      summary: Reset Password
      parameters:
        - in: query
          name: token
          schema:
            type: string
            example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6NDAsImVtYWlsIjoiYiIsImlhdCI6MTcwMTk2MTIwOH0._GJxdj6WMMq3cwXRfR3LVwkSAu74rnPyMdvwhmo2wUU
      requestBody:
        description: change password
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  example: password123
                ConfirmationPassword:
                  type: string
                  example: password213
      responses:
        '200':
          description: change password successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'success'
                  err:
                    type: string
                    example: null
                  data:
                    type: object
                    properties:
                      user:
                        type: object
                        properties:
                          nama:
                            type: string
                            example: John Doe
                          email:
                            type: string
                            example: john@example.com
                          role:
                            type: string
                            example: user
        '400':
          description: new password not same
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: bad request
                  err:
                    type: string
                    example: please ensure that the password and password confirmation match!
                  data:
                    type: string
                    example: null
        '401':
          description: invalid token
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'bad request'
                  err:
                    type: string
                    example: invalid token
                  data:
                    type: string
                    example: null
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  requestBodies:
...
    verify-otp:
      content:
          application/json:
            schema:
              type: object
              properties:
                otp:
                  type: string
                  example: 494387
...